pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adaptive-noise-schedule-diffusion-with-clip-guidance"
version = "0.1.0"
description = "Text-to-image diffusion model with adaptive noise scheduling based on CLIP-measured semantic alignment"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Alireza Shojaei"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "diffusers>=0.25.0",
    "transformers>=4.35.0",
    "accelerate>=0.25.0",
    "peft>=0.7.0",
    "datasets>=2.14.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "tqdm>=4.65.0",
    "PyYAML>=6.0",
    "mlflow>=2.8.0",
    "torchmetrics>=1.2.0",
    "einops>=0.7.0",
    "safetensors>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[tool.setuptools]
packages = ["adaptive_noise_schedule_diffusion_with_clip_guidance"]
package-dir = {"" = "src"}
